added a journal file for the processed virtual file
The journal contain avery elaborated offset, for the stage file, it contains the offset of all data pack wrote into datafile.
In this way last offset wrote in journal is the last data pack for which operation is gone well. The journal is enabled only for the
processed file, not for the destination file or operation.

Signed-off-by: Claudio Bisegni <Claudio.Bisegni@lnf.infn.it>